Rename Findings → AnalyzedDocument; module reshuffle; context placeholder + issue #314
Wraps several coupled renames and moves the earlier "clean up
per-document signatures" PR should have had:

Type renames (public API):
- `Findings` → `AnalyzedDocument`. Symmetric with `Document`
  (input) and `AnonymizedDocument` (output). Pipeline reads:
  Document → analyze_document → AnalyzedDocument →
  anonymize_document → AnonymizedDocument.
- `RawDocument` → `Document`. Renamed schema file too:
  `file/raw.rs` → `file/document.rs`.
- `apply_document` → `anonymize_document`; `ApplyOutcome` →
  `AnonymizedDocument`.

Signature cleanups:
- `Document` gains a `correlation_id: Uuid` field (UUIDv7
  default via `Document::new(bytes, extension)`). Both
  `analyze_document` and `anonymize_document` read the id from
  the document instead of taking a separate parameter.
- `analyze_document` gains a `contexts: &[Context]`
  placeholder. Reserved on the API; no built-in recognizer
  consumes it yet. See #314 for the wiring plan (custom
  recognizer decorator).

Module reshuffle:
- `src/engine/` → `src/pipeline/`. Avoids the redundant
  `engine::Engine` reading; the module is about how the
  pipeline runs, not about the entity called Engine.
- `src/engine/analyzer/` and `src/engine/anonymizer/` promoted
  up one level to `src/analyzer/` and `src/anonymizer/`. They
  hold per-modality plumbing consumed by the pipeline, not
  substructure of it.
- `src/llm/attach.rs` moved into `src/analyzer/llm.rs`. Its
  entry point renamed `attach_lineup` → `attach_llm_lineup`
  to mirror `attach_ner_lineup`.
- `src/analyzer/common.rs`: NER attach code (~60 lines)
  extracted into a new `src/analyzer/ner.rs`. Now `ner.rs`
  and `llm.rs` sit as symmetric peers under `analyzer/`;
  `common.rs` holds only the truly modality-generic helpers
  (pattern, language, dedup).

Visibility:
- `pipeline::analyzed` (holding `AnalyzedDocument`,
  `RecognizedGroup`, `EntityRecord`) is private; types
  re-exported at the crate root.

Observability:
- `COMPONENT` tracing strings renamed `engine::…` →
  `pipeline::…` to match the code structure.

Downstream:
- README badge labels: `rust` / `python` → `runtime` /
  `inference`. The deployment/product framing reads more
  honestly at first glance.

Wire schema unchanged. `Document` is not serialized.
Verified: `make rust-ci` green at defaults and
`--no-default-features`.

Refs #314
