+++ crates/nvisy-engine/src/analyzer/audio.rs
@@ -23,7 +23,8 @@ use elide_stt::{MockBackend as MockSttBackend, SttEnricher};
 use nvisy_core::ner::NerConfig;
 use nvisy_schema::plan::{AnalyzerParams, SttBackendParams, SttEnricherParams};
 
-use super::common::{attach_dedup, attach_ner_lineup, attach_pattern};
+use super::common::{attach_dedup, attach_pattern};
+use super::ner::attach_ner_lineup;
 
 /// Compile `spec` into an audio-modality [`Analyzer`].
 pub(super) fn compile(spec: &AnalyzerParams, ner: &NerConfig) -> Result<Analyzer<Audio>, Error> {
