+++ crates/nvisy-engine/src/analyzer/catalog.rs
