+++ crates/nvisy-engine/src/analyzer/common.rs
@@ -1,14 +1,12 @@
 //! Modality-generic compile helpers shared by every per-modality
 //! analyzer compile function.
 //!
-//! `PatternRecognizer` and `NerRecognizer` are themselves
-//! modality-generic (both impl `Recognizer<M>` for any
-//! `M: TextRecognizable`), so this module hosts the pattern/NER
-//! attach + dedup helpers once and lets the per-modality files
-//! specialise only LLM (which is generic over a different
-//! `LlmModality` trait) and the modality-specific enricher
-//! attachments (`language` for text, `ocr` for image, `stt` for
-//! audio).
+//! `PatternRecognizer` is itself modality-generic (impls
+//! `Recognizer<M>` for any `M: TextRecognizable`), so this
+//! module hosts the pattern-attach helper here alongside the
+//! language enricher for text and the dedup pipeline that runs
+//! after every modality. NER and LLM attach live in
+//! [`super::ner`] and [`super::llm`].
 
 use elide::detection::Analyzer;
 use elide::detection::calibrate::CalibrateLayer;
@@ -18,15 +16,12 @@ use elide::detection::reconcile::tiebreaker::{HighestConfidence, LongestSpan};
 use elide::detection::reconcile::{Merging, ReconcileLayer, Structural};
 use elide::enrichment::lingua::LinguaEnricher;
 use elide::recognition::context::Enhanced;
-use elide::recognition::ner::NerRecognizer;
 use elide::recognition::pattern::PatternRecognizer;
-use elide_bento::BentoNer;
 use elide_core::Error;
 use elide_core::modality::text::Text;
 use elide_core::modality::{Modality, TextRecognizable};
 use elide_core::primitive::ConfidenceThreshold;
 use elide_core::recognition::Recognizer;
-use nvisy_core::ner::{NerBackendConfig, NerConfig, NerRecognizer as ConfigNerRecognizer};
 use nvisy_schema::plan::{
     DeduplicationParams, LanguageEnricherParams, MergingStrategyParams, PatternRecognizerParams,
     TiebreakerParams,
@@ -72,68 +67,6 @@ where
     } else {
         Ok(analyzer.with_recognizer(builder.build()?))
     }
-}
-
-/// Attach every recognizer from the deployment's NER lineup.
-/// Errors when the lineup is empty (compile is only invoked when
-/// the request toggled `ner = true`, so "no recognizers
-/// configured" is user-visible). Modality-generic for any
-/// `M: TextRecognizable`.
-pub(super) fn attach_ner_lineup<M>(
-    mut analyzer: Analyzer<M>,
-    ner: &NerConfig,
-) -> Result<Analyzer<M>, Error>
-where
-    M: TextRecognizable,
-    NerRecognizer: Recognizer<M> + 'static,
-{
-    if ner.recognizers.is_empty() {
-        return Err(Error::new(
-            elide_core::ErrorKind::Validation,
-            "AnalyzerParams.recognizers.ner = true but the deployment has no NER \
-             recognizer configured; add one to `[[ner.recognizers]]` in the \
-             deployment config or leave `ner = false`",
-        ));
-    }
-    for recognizer in &ner.recognizers {
-        analyzer = attach_ner_one(analyzer, recognizer)?;
-    }
-    Ok(analyzer)
-}
-
-fn attach_ner_one<M>(
-    analyzer: Analyzer<M>,
-    spec: &ConfigNerRecognizer,
-) -> Result<Analyzer<M>, Error>
-where
-    M: TextRecognizable,
-    NerRecognizer: Recognizer<M> + 'static,
-{
-    let mut builder = NerRecognizer::builder().with_name(spec.name.clone());
-    match &spec.backend {
-        NerBackendConfig::Bento { base_url, model } => {
-            builder = builder.with_backend(BentoNer::new(base_url.clone(), model.clone())?);
-        }
-        #[cfg(feature = "test-utils")]
-        NerBackendConfig::Mock => {
-            builder = builder.with_mock_backend();
-        }
-        // `NerBackendConfig` is `#[non_exhaustive]`. A future
-        // variant reaching this arm should surface as a
-        // Validation error rather than silently dropping the
-        // recognizer.
-        _ => {
-            return Err(Error::new(
-                elide_core::ErrorKind::Validation,
-                format!(
-                    "NER recognizer `{}` uses a backend kind this engine binary \
-                     doesn't understand; upgrade the engine or downgrade the config",
-                    spec.name,
-                ),
-            ));
-        }
-    }
-    Ok(analyzer.with_recognizer(builder.build()?))
 }
 
 /// Append the deduplication layers: calibrate → reconcile
