+++ crates/nvisy-engine/src/anonymizer/image.rs
